Spring Boot - Adicionando Detalhes do Evento

+++ src/main/java/com/eventoapp/eventoapp/models/Evento.java
@@ -19,7 +19,7 @@ public class Evento implements Serializable {
     private String local;
     private String data;
 
-    public long getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
